Don't define classes on default package
