--- detekt-rules-style/src/test/resources/java/ObjectLiteralToLambda.java
@@ -1,15 +1,0 @@
-public interface ObjectLiteralToLambda {
-
-    @FunctionalInterface
-    interface SamWithDefaultMethods {
-        void foo();
-        default void bar() {}
-        default void baz() {}
-    }
-
-    interface OnlyDefaultMethods {
-        default void foo() {}
-        default void bar() {}
-        default void baz() {}
-    }
-}
