+++ detekt-rules-style/src/test/resources/java/com/example/fromjava/OnlyDefaultMethods.java
@@ -1,0 +1,7 @@
+package com.example.fromjava;
+
+interface OnlyDefaultMethods {
+    default void foo() {}
+    default void bar() {}
+    default void baz() {}
+}
