+++ detekt-rules-style/src/test/resources/java/com/example/fromjava/SamWithDefaultMethods.java
@@ -1,0 +1,8 @@
+package com.example.fromjava;
+
+@FunctionalInterface
+interface SamWithDefaultMethods {
+    void foo();
+    default void bar() {}
+    default void baz() {}
+}
